refactor(cw3-flex-msig): Move contract tests into separate module
