+++ contracts/core-cw3-flex-msig/Cargo.toml
@@ -35,6 +35,7 @@ cw4-group = { workspace = true }
 cw-multi-test = { workspace = true }
 cw20-base = { workspace = true }
 easy-addr = { workspace = true }
+anyhow = { workspace = true }
 
 [[bin]]
 name = "cw3-flex-multisig-schema"
